conf.h: separate interface and implementation
