+++ src/common/curl_tools.c
@@ -21,6 +21,7 @@
 
 #include "common/darktable.h"
 #include "common/curl_tools.h"
+#include "common/file_location.h"
 #include "control/control.h"
 
 void dt_curl_init(CURL *curl, gboolean verbose)
