+++ src/common/image.c
@@ -21,6 +21,7 @@
 #include "common/darktable.h"
 #include "common/debug.h"
 #include "common/exif.h"
+#include "common/file_location.h"
 #include "common/grouping.h"
 #include "common/history.h"
 #include "common/image_cache.h"
