+++ src/gui/gtk.c
@@ -22,6 +22,7 @@
 #endif
 #include "common/collection.h"
 #include "common/colorspaces.h"
+#include "common/file_location.h"
 #include "common/image.h"
 #include "common/image_cache.h"
 #include "develop/develop.h"
