+++ src/imageio/storage/email.c
@@ -17,6 +17,7 @@
 */
 
 #include "common/darktable.h"
+#include "common/file_location.h"
 #include "common/image.h"
 #include "common/image_cache.h"
 #include "common/imageio.h"
