+++ src/iop/colorout.c
@@ -22,6 +22,7 @@
 #include "bauhaus/bauhaus.h"
 #include "common/colorspaces.h"
 #include "common/colorspaces_inline_conversions.h"
+#include "common/file_location.h"
 #include "common/opencl.h"
 #include "control/conf.h"
 #include "control/control.h"
