+++ src/libs/export.c
@@ -21,6 +21,7 @@
 #include "common/colorspaces.h"
 #include "common/darktable.h"
 #include "common/debug.h"
+#include "common/file_location.h"
 #include "common/imageio_module.h"
 #include "common/styles.h"
 #include "control/conf.h"
