+++ src/libs/geotagging.c
@@ -18,6 +18,7 @@
 // #include "common/darktable.h"
 // #include "control/control.h"
 #include "common/debug.h"
+#include "common/file_location.h"
 #include "common/image_cache.h"
 #include "control/conf.h"
 // #include "gui/accelerators.h"
