+++ src/libs/modulelist.c
@@ -20,6 +20,7 @@
 
 #include "common/darktable.h"
 #include "common/debug.h"
+#include "common/file_location.h"
 #include "common/image_cache.h"
 #include "control/conf.h"
 #include "control/control.h"
